requirements.txt
pyyaml
youtube-transcript-api
openai
tiktoken
orjson
chromadb
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0